#4281 move add/delete ruleset type to admin tools
