+++ src/frontend/src/pages/AdminToolsPage/AdminToolsProjectsConfig.tsx
@@ -12,6 +12,7 @@ import ConfluenceLink from './ProjectsConfig/ConfluenceLink';
 import PartReviewSampleImage from './ProjectsConfig/PartReviewSampleImage';
 import CommonMistakesTable from './ProjectsConfig/CommonMistakesTable';
 import PartTagsTable from './ProjectsConfig/PartTagsTable';
+import RulesetTypesTable from './ProjectsConfig/RulesetTypesTable';
 
 const AdminToolsProjectsConfig: React.FC = () => {
   return (
@@ -35,6 +36,10 @@ const AdminToolsProjectsConfig: React.FC = () => {
         Cars Config
       </Typography>
       <CarsTable />
+      <Typography variant="h5" gutterBottom borderBottom={1} color="#ef4345" borderColor={'white'}>
+        Rules Config
+      </Typography>
+      <RulesetTypesTable />
       <Typography variant="h5" gutterBottom borderBottom={1} color="#ef4345" borderColor={'white'}>
         Links Config
       </Typography>
