+++ src/frontend/src/pages/RulesPage/RulesetTypePage.tsx
@@ -2,24 +2,8 @@ import React from 'react';
 import PageLayout from '../../components/PageLayout';
 import { Box } from '@mui/material';
 import RulesetTypeTable from './components/RulesetTypeTable';
-import { NERButton } from '../../components/NERButton';
-import AddRulesetTypeModal from './components/AddRulesetTypeModal';
-import { useState } from 'react';
-import { useCreateRulesetType } from '../../hooks/rules.hooks';
 
 const RulesetTypePage: React.FC = () => {
-  const [addRulesetTypeModalShow, setAddRulesetTypeModalShow] = useState(false);
-
-  const { mutateAsync: createRulesetType } = useCreateRulesetType();
-
-  const handleAddRulesetTypeConfirm = async (data: { name: string }) => {
-    await createRulesetType({ name: data.name });
-  };
-
-  const handleAddRulesetTypeCancel = () => {
-    setAddRulesetTypeModalShow(false);
-  };
-
   return (
     <>
       <PageLayout title="Ruleset Types">
@@ -46,27 +30,6 @@ const RulesetTypePage: React.FC = () => {
                 ml: '20px'
               })}
             />
-            <Box
-              sx={{
-                display: 'flex',
-                justifyContent: { xs: 'center', md: 'flex-end' },
-                pr: '30px',
-                pb: 2
-              }}
-            >
-              <NERButton
-                variant="contained"
-                sx={{ color: '#ededed' }}
-                onClick={() => setAddRulesetTypeModalShow(!addRulesetTypeModalShow)}
-              >
-                Add Ruleset Type
-              </NERButton>
-              <AddRulesetTypeModal
-                open={addRulesetTypeModalShow}
-                onHide={handleAddRulesetTypeCancel}
-                onFormSubmit={handleAddRulesetTypeConfirm}
-              />
-            </Box>
           </Box>
         </Box>
       </PageLayout>
