+++ src/frontend/src/pages/RulesPage/components/RulesetTypeTable.tsx
@@ -12,8 +12,7 @@ import {
   Card,
   CardContent,
   Typography,
-  Stack,
-  IconButton
+  Stack
 } from '@mui/material';
 import { useHistory } from 'react-router-dom';
 import { datePipe } from '../../../utils/pipes';
@@ -24,9 +23,6 @@ import ErrorPage from '../../ErrorPage';
 import { RulesetType } from 'shared';
 import { NERButton } from '../../../components/NERButton';
 import { useToast } from '../../../hooks/toasts.hooks';
-import { useState } from 'react';
-import RulesetTypeDeleteModal from './RulesetTypeDeleteModal';
-import { Delete } from '@mui/icons-material';
 
 type RulesetTypeColumnId = 'id' | 'name' | 'lastUpdated' | 'revisions' | 'actions' | 'delete';
 
@@ -35,20 +31,12 @@ interface RulesetTypeHeadCell {
   label: string;
 }
 
-interface RulesetTypeDeleteButtonProps {
-  rulesetTypeId: string;
-  name: string;
-  onDelete: (rulesetTypeId: string, name: string) => void;
-}
-
 const RulesetTypeTable: React.FC = () => {
   const history = useHistory();
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down('md'));
-  const toast = useToast();
 
   const { data: rulesetTypes = [], isLoading, error } = useAllRulesetTypes();
-  const { mutateAsync: deleteRulesetType } = useDeleteRulesetType();
 
   const headCells: readonly RulesetTypeHeadCell[] = [
     {
@@ -66,56 +54,11 @@ const RulesetTypeTable: React.FC = () => {
     {
       id: 'actions',
       label: 'Actions'
-    },
-    {
-      id: 'delete',
-      label: ''
     }
   ];
 
   const handleViewRulesetType = (rulesetTypeId: string) => {
     history.push(routes.RULESET_BY_ID.replace(':rulesetTypeId', rulesetTypeId));
-  };
-
-  const handleDeleteRulesetType = async (rulesetTypeId: string, name: string) => {
-    const rulesetType = rulesetTypes.find((rt) => rt.rulesetTypeId === rulesetTypeId);
-    if (rulesetType && rulesetType.revisionFiles.length > 0) {
-      toast.error('Cannot delete ruleset type with existing revisions');
-      return;
-    }
-
-    try {
-      await deleteRulesetType(rulesetTypeId);
-      toast.success(`Ruleset Type: ${name} deleted successfully!`);
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        toast.error(error.message);
-      }
-    }
-  };
-
-  const RulesetTypeDeleteButton: React.FC<RulesetTypeDeleteButtonProps> = ({ rulesetTypeId, name, onDelete }) => {
-    const [showDeleteModal, setShowDeleteModal] = useState(false);
-
-    const handleDeleteSubmit = () => {
-      onDelete(rulesetTypeId, name);
-      setShowDeleteModal(false);
-    };
-
-    return (
-      <>
-        <IconButton type="button" sx={{ mx: 1 }} onClick={() => setShowDeleteModal(true)}>
-          <Delete />
-        </IconButton>
-        {showDeleteModal && (
-          <RulesetTypeDeleteModal
-            rulesetTypeName={name}
-            onDelete={handleDeleteSubmit}
-            onHide={() => setShowDeleteModal(false)}
-          />
-        )}
-      </>
-    );
   };
 
   if (error) return <ErrorPage message={error.message} />;
@@ -163,8 +106,6 @@ const RulesetTypeTable: React.FC = () => {
                         '&:hover': {
                           backgroundColor: theme.palette.grey[700]
                         },
-                        marginRight: '10px',
-                        padding: '4px',
                         lineHeight: 1,
                         borderRadius: '6px'
                       }}
@@ -172,11 +113,6 @@ const RulesetTypeTable: React.FC = () => {
                     >
                       View Rulesets
                     </NERButton>
-                    <RulesetTypeDeleteButton
-                      rulesetTypeId={rulesetType.rulesetTypeId}
-                      name={rulesetType.name}
-                      onDelete={handleDeleteRulesetType}
-                    />
                   </Box>
                 </Box>
               </CardContent>
@@ -228,8 +164,6 @@ const RulesetTypeTable: React.FC = () => {
                           '&:hover': {
                             backgroundColor: theme.palette.grey[700]
                           },
-                          marginRight: '10px',
-                          padding: '4px',
                           lineHeight: 1,
                           borderRadius: '6px'
                         }}
@@ -237,13 +171,6 @@ const RulesetTypeTable: React.FC = () => {
                       >
                         View Rulesets
                       </NERButton>
-                    </TableCell>
-                    <TableCell align="center" sx={{ width: '60px', paddingLeft: '0px' }}>
-                      <RulesetTypeDeleteButton
-                        rulesetTypeId={rulesetType.rulesetTypeId}
-                        name={rulesetType.name}
-                        onDelete={handleDeleteRulesetType}
-                      />
                     </TableCell>
                   </TableRow>
                 ))
